PROJECT/back/requirements.txt

backoff==1.10.0
beautifulsoup4==4.9.3
bidict==0.21.2
billiard==3.6.4.0
blinker==1.4
boto3==1.24.80
botocore==1.27.80
CacheControl==0.12.11
cachelib==0.1.1
cachetools==5.0.0
celery==4.4.7
certifi==2021.10.8
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.9
ci-info==0.3.0
Click==7.0
colorama==0.4.4
comtypes==1.1.14
configobj==5.0.8
configparser==5.3.0
cryptography==3.4.7
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
decouple==0.0.7
defusedxml==0.7.1
Deprecated==1.2.12
distlib==0.3.6
dj-database-url==0.5.0
Django==3.0
django-allauth==0.39.1
django-cors-headers==3.7.0
django-crispy-forms==1.9.2
django-environ==0.8.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-rest-auth==0.9.5
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.6.0
entrypoints==0.4
et-xmlfile==1.1.0
etelemetry==0.3.0
executing==0.10.0
face-recognition-models==0.3.0
Faker==2.0.1
filelock==3.12.0
firebase-admin==6.0.1
fitz==0.0.1.dev2
flasgger==0.9.5
flatbuffers==23.5.26
fonttools==4.28.5
freezegun==1.0.0
frontend==0.0.3
frozenlist==1.3.3
future==0.18.3
fuzzywuzzy==0.18.0
gast==0.4.0
google==3.0.0
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-firestore==2.5.3
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.0
grpcio==1.50.0
grpcio-status==1.48.2
gunicorn==20.0.4
h11==0.14.0
h5py==3.7.0
honcho==1.0.1
html5lib==1.0b10
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.17.1
ipython==8.4.0
isodate==0.6.1
itsdangerous==1.1.0
jax==0.4.10
jedi==0.18.1
jet==1.0
Jinja2==2.10.1
jmespath==1.0.1
joblib==1.1.0
jsmin==2.2.2
jsonpickle==1.2
jsonschema==3.2.0
jupyter-client==7.4.6
jupyter-core==5.0.0
keras==2.12.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
kombu==4.6.11
lassie==0.11.9
Levenshtein==0.21.0
libclang==14.0.6
librosa==0.9.2
llvmlite==0.39.0
logzero==1.7.0
looseversion==1.1.2
lxml==4.6.3
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.24.3
mathparse==0.1.2
matplotlib==3.5.1
matplotlib-inline==0.1.6
mistune==0.8.4
ml-dtypes==0.1.0
mock==4.0.2
MouseInfo==0.1.3
msgpack==1.0.4
multidict==6.0.4
mysql-connector==2.2.9
mysqlclient==2.0.3
nest-asyncio==1.5.6
networkx==3.0
nibabel==5.0.1
nipype==1.8.5
nltk==3.7
numba==0.56.0
numpy==1.23.5
oauthlib==3.2.0
openai==0.27.8
opencv-python==4.6.0.66
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.2
parso==0.8.3
pathlib==1.0.1
pdfminer.six==20200517
pdfplumber==0.5.28
pdfquery==0.4.3
pickleshare==0.7.5
Pillow==8.3.2
pipenv==2023.6.2
pipupgrade==1.5.1
platformdirs==3.5.1
playsound==1.3.0
pooch==1.6.0
prettytable==2.1.0
prompt-toolkit==3.0.30
proto-plus==1.22.1
protobuf==3.19.6
prov==2.0.0
psutil==5.9.4
psycopg2==2.8.6
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.10.1
pydot==1.4.2
PyGetWindow==0.0.9
Pygments==2.6.1
PyJWT==2.6.0
PyMsgBox==1.0.9
PyMuPDF==1.21.1
PyMySQL==1.0.2
pyotp==2.6.0
pyparsing==3.0.8
pyPdf==1.13
PyPDF2==1.26.0
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
PyQRCode==1.2.1
pyquery==2.0.0
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.26
python-dateutil==2.8.2
python-dotenv==0.15.0
python-editor==1.0.4
python-engineio==4.1.0
python-Levenshtein==0.21.0
python-oembed==0.2.4
python-socketio==5.2.1
python3-openid==3.2.0
pyttsx3==2.90
PyTweening==1.0.3
pytz==2021.1
pywhatkit==3.9
pywin32==305
pyxnat==1.5
PyYAML==5.4.1
pyzmq==24.0.1
qrcode==7.3.1
rapidfuzz==3.1.1
raygun4py==4.3.0
rdflib==6.2.0
readability-lxml==0.8.1
redis==3.3.8
regex==2022.10.31
reportlab==3.6.12
requests==2.26.0
requests-oauthlib==1.3.1
resampy==0.4.0
roman==4.0
rq==1.1.0
rsa==4.8
s3transfer==0.6.0
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.12.2
simplejson==3.18.3
six==1.16.0
sklearn==0.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.5
SoundFile==0.10.3.post1
soupsieve==2.0.1
SpeechRecognition==3.8.1
SQLAlchemy==1.3.8
sqlparse==0.4.2
stack-data==0.4.0
starlette==0.25.0
stripe==2.56.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
text-unidecode==1.2
tf==1.0.0
threadpoolctl==3.1.0
toml==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
traits==6.3.2
twilio==6.57.0
typing-extensions==4.3.0
tzdata==2021.5
uritemplate==4.1.1
url-metadata==0.1.6
urllib3==1.26.7
uvicorn==0.20.0
vine==1.3.0
virtualenv==20.23.0
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.6
vonage==2.5.5
Wand==0.6.7
wcwidth==0.2.5
webassets==0.12.1
webencodings==0.5.1
Werkzeug==2.2.2
whitenoise==5.2.0
wikipedia==1.4.0
wrapt==1.11.2
WTForms==2.2.1
xlrd==2.0.1
yarl==1.9.2
zipp==3.8.1
